+++ src/fetch/SecFetchFileOutputCache.ts
@@ -6,9 +6,11 @@
 
 import { TaskInput, TaskOutput, TaskOutputRepository } from "@workglow/task-graph";
 import { FetchUrlTaskOutput } from "@workglow/tasks";
+import { globalServiceRegistry } from "@workglow/util";
 import { mkdirSync } from "node:fs";
 import { mkdir, readFile, writeFile, stat } from "node:fs/promises";
 import path from "node:path";
+import { SEC_DRY_RUN } from "../config/tokens";
 import { secDate } from "../util/parseDate";
 import { YYYYdMMdDD } from "../util/parseDate";
 
@@ -66,6 +68,9 @@ export class SecFetchFileOutputCache extends TaskOutputRepository {
    * @param output The task output to save
    */
   async saveOutput(taskType: string, input: TaskInput, output: TaskOutput): Promise<void> {
+    if (globalServiceRegistry.has(SEC_DRY_RUN) && globalServiceRegistry.get(SEC_DRY_RUN)) {
+      return;
+    }
     const filePath = path.join(this.folderPath, this.inputToFileName(input));
     await mkdir(path.dirname(filePath), { recursive: true });
     await writeFile(filePath, this.outputSerializer(output, input.response_type as string), {
